microservices/auth_service/requirements.txt
Flask==3.0.3
PyJWT==2.8.0
gunicorn==22.0.0
python-dotenv==1.0.1
